requirements.txt
anthropic
requests
python-dotenv
flask
Pillow
